pyproject.toml
[project]
name = "sbt"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ning0426", email = "15996204@qq.com" }
]
dependencies = [
    "backtesting>=0.6.5",
    "xtquant>=250516.1.1",
    "click>=8.2.1",
    "python-dotenv>=1.1.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "duckdb>=1.3.2",
    "polars>=1.32.2",
    "ipykernel>=6.30.1",
    "pyarrow>=21.0.0",
    "ta-lib>=0.6.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
sbt = "sbt.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sbt"]
